requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
pydantic
pydantic-settings
psycopg2-binary
psycopg2
